// FatmaEvent_Soutenance-main/coreui-free-angular-admin-template-main/src/app/models/soutenance.model.ts
// src/app/models/soutenance.model.ts
export interface Soutenance {
  idSoutenance?: number; // Optional because it's auto-generated by the backend
  idEnseignant: number;
  idEtudiant: number;
  dateSoutenance: string; // Use string for simplicity (format: "YYYY-MM-DD")
  lieu: string;
  salle: string;
  time: string; // Use string for simplicity (format: "HH:MM:SS")
  bloc: string;
}